+++ hutool-core/src/main/java/cn/hutool/core/bean/BeanUtil.java
@@ -9,29 +9,12 @@ import cn.hutool.core.convert.Convert;
 import cn.hutool.core.lang.Editor;
 import cn.hutool.core.map.CaseInsensitiveMap;
 import cn.hutool.core.map.MapUtil;
-import cn.hutool.core.util.ArrayUtil;
-import cn.hutool.core.util.ClassUtil;
-import cn.hutool.core.util.ModifierUtil;
-import cn.hutool.core.util.ObjectUtil;
-import cn.hutool.core.util.ReflectUtil;
-import cn.hutool.core.util.StrUtil;
+import cn.hutool.core.util.*;
 
-import java.beans.BeanInfo;
-import java.beans.IntrospectionException;
-import java.beans.Introspector;
-import java.beans.PropertyDescriptor;
-import java.beans.PropertyEditor;
-import java.beans.PropertyEditorManager;
+import java.beans.*;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
+import java.util.*;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
@@ -116,7 +99,7 @@ public class BeanUtil {
 				if (method.getParameterCount() == 0) {
 					final String name = method.getName();
 					if (name.startsWith("get") || name.startsWith("is")) {
-						if(false == "getClass".equals(name)){
+						if (false == "getClass".equals(name)) {
 							return true;
 						}
 					}
@@ -309,24 +292,32 @@ public class BeanUtil {
 
 	/**
 	 * 设置字段值，通过反射设置字段值，并不调用setXXX方法<br>
-	 * 对象同样支持Map类型，fieldNameOrIndex即为key
+	 * 对象同样支持Map类型，fieldNameOrIndex即为key，支持：
+	 * <ul>
+	 *     <li>Map</li>
+	 *     <li>List</li>
+	 *     <li>Bean</li>
+	 * </ul>
 	 *
 	 * @param bean             Bean
 	 * @param fieldNameOrIndex 字段名或序号，序号支持负数
 	 * @param value            值
+	 * @return bean，当为数组时，返回一个新的数组
 	 */
 	@SuppressWarnings({"unchecked", "rawtypes"})
-	public static void setFieldValue(Object bean, String fieldNameOrIndex, Object value) {
+	public static Object setFieldValue(Object bean, String fieldNameOrIndex, Object value) {
 		if (bean instanceof Map) {
 			((Map) bean).put(fieldNameOrIndex, value);
 		} else if (bean instanceof List) {
 			ListUtil.setOrPadding((List) bean, Convert.toInt(fieldNameOrIndex), value);
 		} else if (ArrayUtil.isArray(bean)) {
-			ArrayUtil.setOrAppend(bean, Convert.toInt(fieldNameOrIndex), value);
+			// issue#3008，追加产生新数组，此处返回新数组
+			return ArrayUtil.setOrAppend(bean, Convert.toInt(fieldNameOrIndex), value);
 		} else {
 			// 普通Bean对象
 			ReflectUtil.setFieldValue(bean, fieldNameOrIndex, value);
 		}
+		return bean;
 	}
 
 	/**
@@ -611,6 +602,7 @@ public class BeanUtil {
 	}
 
 	// --------------------------------------------------------------------------------------------- beanToMap
+
 	/**
 	 * 将bean的部分属性转换成map<br>
 	 * 可选拷贝哪些属性值，默认是不忽略值为{@code null}的值的。
@@ -623,7 +615,7 @@ public class BeanUtil {
 	public static Map<String, Object> beanToMap(Object bean, String... properties) {
 		int mapSize = 16;
 		Editor<String> keyEditor = null;
-		if(ArrayUtil.isNotEmpty(properties)){
+		if (ArrayUtil.isNotEmpty(properties)) {
 			mapSize = properties.length;
 			final Set<String> propertiesSet = CollUtil.set(false, properties);
 			keyEditor = property -> propertiesSet.contains(property) ? property : null;
@@ -733,7 +725,7 @@ public class BeanUtil {
 	 * @return 目标对象
 	 */
 	public static <T> T copyProperties(Object source, Class<T> tClass, String... ignoreProperties) {
-		if(null == source){
+		if (null == source) {
 			return null;
 		}
 		T target = ReflectUtil.newInstanceIfPossible(tClass);
@@ -773,7 +765,7 @@ public class BeanUtil {
 	 * @param copyOptions 拷贝选项，见 {@link CopyOptions}
 	 */
 	public static void copyProperties(Object source, Object target, CopyOptions copyOptions) {
-		if(null == source){
+		if (null == source) {
 			return;
 		}
 		BeanCopier.create(source, target, ObjectUtil.defaultIfNull(copyOptions, CopyOptions::create)).copy();
@@ -980,14 +972,14 @@ public class BeanUtil {
 	/**
 	 * 判断source与target的所有公共字段的值是否相同
 	 *
-	 * @param source 待检测对象1
-	 * @param target 待检测对象2
+	 * @param source           待检测对象1
+	 * @param target           待检测对象2
 	 * @param ignoreProperties 不需要检测的字段
 	 * @return 判断结果，如果为true则证明所有字段的值都相同
-	 * @since 5.8.4
 	 * @author Takak11
+	 * @since 5.8.4
 	 */
-	public static boolean isCommonFieldsEqual(Object source, Object target, String...ignoreProperties) {
+	public static boolean isCommonFieldsEqual(Object source, Object target, String... ignoreProperties) {
 
 		if (null == source && null == target) {
 			return true;
@@ -1003,7 +995,7 @@ public class BeanUtil {
 		sourceFields.removeAll(Arrays.asList(ignoreProperties));
 
 		for (String field : sourceFields) {
-			if(ObjectUtil.notEqual(sourceFieldsMap.get(field), targetFieldsMap.get(field))){
+			if (ObjectUtil.notEqual(sourceFieldsMap.get(field), targetFieldsMap.get(field))) {
 				return false;
 			}
 		}